perf(web): lazy-load Mermaid/Streamdown and graph viewer off run-load path
The run-detail page eagerly bundled ~2.3 MB of JS that the trace view never needs: Streamdown (which transitively bundles Mermaid + Shiki, ~1.7 MB) via the rarely-rendered conversation view, and the flow-graph viewer (~600 KB) behind the currently-hidden Graph tab. Both sat on the initial-load critical path, blocking first render of the trace.

Lazy-load both (React.lazy + Suspense) so they are code-split and fetched only when actually needed. Measured against a 603-event run, this cuts run-page JS from 3418 KB to 2183 KB (-36%) and removes Mermaid's parser/regex init from the load path.
